moved encryption logic into seperate file
