+++ plugin/init.lua
@@ -52,113 +52,6 @@ local function get_file_path(file_name, type, opt_name)
 	return string.format("%s%s" .. separator .. "%s.json", pub.save_state_dir, type, file_name:gsub(separator, "+"))
 end
 
----executes cmd and passes input to stdin
----@param cmd string command to be run
----@param input string input to stdin
----@return boolean
----@return string
-local function execute_cmd_with_stdin(cmd, input)
-	if is_windows and #input < 32000 then -- Check if input is larger than max cmd length on Windows
-		cmd = string.format("%s | %s", wezterm.shell_join_args({ "Write-Output", "-NoEnumerate", input }), cmd)
-		local process_args = { "pwsh.exe", "-NoProfile", "-Command", cmd }
-
-		local success, stdout, stderr = wezterm.run_child_process(process_args)
-		if success then
-			return success, stdout
-		else
-			return success, stderr
-		end
-	elseif #input < 150000 and not is_windows then -- Check if input is larger than common max on MacOS and Linux
-		cmd = string.format("%s | %s", wezterm.shell_join_args({ "echo", "-E", "-n", input }), cmd)
-		local process_args = { os.getenv("SHELL"), "-c", cmd }
-
-		local success, stdout, stderr = wezterm.run_child_process(process_args)
-		if success then
-			return success, stdout
-		else
-			return success, stderr
-		end
-	else
-		-- redirect stderr to stdout to test if cmd will execute
-		-- can't check on Windows because it doesn't support /dev/stdin
-		if not is_windows then
-			local stdout = io.popen(cmd .. " 2>&1", "r")
-			if not stdout then
-				return false, "Failed to execute: " .. cmd
-			end
-			local stderr = stdout:read("*all")
-			stdout:close()
-			if stderr ~= "" then
-				wezterm.log_error(stderr)
-				return false, stderr
-			end
-		end
-		-- if no errors, execute cmd using stdin with input
-		local stdin = io.popen(cmd, "w")
-		if not stdin then
-			return false, "Failed to execute: " .. cmd
-		end
-		stdin:write(input)
-		stdin:flush()
-		stdin:close()
-		return true, '"' .. cmd .. '" <input> ran successfully.'
-	end
-end
-
----@alias encryption_opts {enable: boolean, method: string, private_key: string | nil, public_key: string | nil, encrypt: fun(file_path: string, lines: string), decrypt: fun(file_path: string): string}
-pub.encryption = {
-	enable = false,
-	method = "age",
-	private_key = nil,
-	public_key = nil,
-	encrypt = function(file_path, lines)
-		local cmd = string.format(
-			"%s -r %s -o %s",
-			pub.encryption.method,
-			pub.encryption.public_key,
-			file_path:gsub(" ", "\\ ")
-		)
-
-		if pub.encryption.method:find("gpg") then
-			cmd = string.format(
-				"%s --batch --yes --encrypt --recipient %s --output %s",
-				pub.encryption.method,
-				pub.encryption.public_key,
-				file_path:gsub(" ", "\\ ")
-			)
-		end
-
-		local success, output = execute_cmd_with_stdin(cmd, lines)
-		if not success then
-			error("Encryption failed:" .. output)
-		end
-	end,
-	decrypt = function(file_path)
-		local cmd = { pub.encryption.method, "-d", "-i", pub.encryption.private_key, file_path }
-
-		if pub.encryption.method:find("gpg") then
-			cmd = { pub.encryption.method, "--batch", "--yes", "--decrypt", file_path }
-		end
-
-		local success, stdout, stderr = wezterm.run_child_process(cmd)
-		if not success then
-			error("Decryption failed: " .. stderr)
-		end
-
-		return stdout
-	end,
-}
-
---- Merges user-supplied options with default options
---- @param user_opts encryption_opts
-function pub.set_encryption(user_opts)
-	for k, v in pairs(user_opts) do
-		if v ~= nil then
-			pub.encryption[k] = v
-		end
-	end
-end
-
 --- Sanitize the input by replacing control characters and invalid UTF-8 sequences with valid \uxxxx unicode
 --- @param data string
 --- @return string
@@ -185,9 +78,8 @@ local function write_state(file_path, state, event_type)
 			return pub.encryption.encrypt(file_path, json_state)
 		end)
 		if not ok then
-			wezterm.log_error("Encryption failed: ")
-			wezterm.log_error(err)
-			wezterm.emit("resurrect.error", err)
+			wezterm.emit("resurrect.error", "Encryption failed: " .. tostring(err))
+			wezterm.log_error("Decryption failed: " .. tostring(err))
 		else
 			wezterm.emit("resurrect.encrypt.finished", file_path)
 		end
@@ -362,11 +254,18 @@ function pub.delete_state(file_path)
 	wezterm.emit("resurrect.delete_state.finished", file_path)
 end
 
+--- Merges user-supplied options with default options
+--- @param user_opts encryption_opts
+function pub.set_encryption(user_opts)
+	require("plugin.resurrect.encryption").set_encryption(user_opts)
+end
+
 -- Export submodules
 pub.workspace_state = require("resurrect.workspace_state")
 pub.window_state = require("resurrect.window_state")
 pub.tab_state = require("resurrect.tab_state")
 pub.fuzzy_loader = require("resurrect.fuzzy_loader")
+pub.encryption = require("plugin.resurrect.encryption").encryption
 
 ---Changes the directory to save the state to
 ---@param directory string
